requirements.txt
python-dateutil==2.8.2
tabulate==0.9.0
colorama==0.4.6